select of category is moved to separate component.
